refactor: Rename internal/visitor to lib/visitor
