+++ internal/addon/rules/fileNamesLowerSnakeCaseRule.go
@@ -7,7 +7,7 @@ import (
 	"github.com/yoheimuta/protolint/internal/stringsutil"
 
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/internal/visitor"
+	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
 	"github.com/yoheimuta/protolint/internal/strs"
 )
