+++ internal/addon/rules/lib/visitor/baseAddVisitor.go
