+++ internal/addon/rules/lib/visitor/hasExtendedVisitor.go
