+++ internal/addon/rules/lib/visitor/hasExtendedVisitor_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/yoheimuta/go-protoparser/parser/meta"
 
-	"github.com/yoheimuta/protolint/internal/addon/rules/internal/visitor"
+	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 
 	"github.com/yoheimuta/protolint/internal/linter/report"
