+++ internal/addon/rules/lib/visitor/internal/disablerule/command.go
