+++ internal/addon/rules/lib/visitor/internal/disablerule/commands.go
